Gateway: Create package core with module database
