+++ gateway/core/database.py
@@ -1,0 +1,14 @@
+import os
+from typing import Annotated
+from fastapi import Depends
+from sqlmodel import create_engine, Session
+
+engine = create_engine(os.getenv("DATABASE_URL"))
+
+
+def get_session():
+    with Session(engine) as session:
+        yield session
+
+
+SessionDep = Annotated[Session, Depends(get_session)]
